requirements.txt
mediafire>=0.6.1
requests>=2.28.0
python-dotenv>=1.0.0
